Fix issue of improper NSInteger casting

+++ CTAssetsPickerController/CTAssetsGridSelectedView.m
@@ -155,7 +155,7 @@
 - (void)setSelectionIndex:(NSUInteger)selectionIndex;
 {
     _selectionIndex = selectionIndex;
-    self.selectionIndexLabel.text = [NSString stringWithFormat:@"%lu", selectionIndex + 1];
+    self.selectionIndexLabel.text = [NSString stringWithFormat:@"%lu", (unsigned long)selectionIndex + 1];
 }
